Move PrepareData to ByteArrayExtensions
