+++ src/lib/Sander0542.CMLedController/LedControllerDevice.cs
@@ -13,12 +13,12 @@ namespace Sander0542.CMLedController
 
         protected override async Task WriteAsync(byte[] data, CancellationToken token = default)
         {
-            await Device.WriteAsync(PrepareData(data), token);
+            await Device.WriteAsync(data, token);
         }
 
         protected override async Task<byte[]> WriteAndReadAsync(byte[] data, CancellationToken token = default)
         {
-            return (await Device.WriteAndReadAsync(PrepareData(data), token)).Data;
+            return (await Device.WriteAndReadAsync(data, token)).Data;
         }
 
         public override void Dispose()
